+++ tests/starknet/keccak.rs
@@ -18,7 +18,7 @@ impl StarkNetSyscallHandler for SyscallHandler {
     fn get_block_hash(
         &self,
         block_number: u64,
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<cairo_felt::Felt252> {
         println!("Called `get_block_hash({block_number})` from MLIR.");
         Ok(Felt252::from_bytes_be(b"get_block_hash ok"))
@@ -26,7 +26,7 @@ impl StarkNetSyscallHandler for SyscallHandler {
 
     fn get_execution_info(
         &self,
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<cairo_native::starknet::ExecutionInfo> {
         println!("Called `get_execution_info()` from MLIR.");
         Ok(ExecutionInfo {
@@ -56,7 +56,7 @@ impl StarkNetSyscallHandler for SyscallHandler {
         contract_address_salt: cairo_felt::Felt252,
         calldata: &[cairo_felt::Felt252],
         deploy_from_zero: bool,
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<(cairo_felt::Felt252, Vec<cairo_felt::Felt252>)> {
         println!("Called `deploy({class_hash}, {contract_address_salt}, {calldata:?}, {deploy_from_zero})` from MLIR.");
         Ok((
@@ -65,7 +65,7 @@ impl StarkNetSyscallHandler for SyscallHandler {
         ))
     }
 
-    fn replace_class(&self, class_hash: cairo_felt::Felt252, _gas: &mut u64) -> SyscallResult<()> {
+    fn replace_class(&self, class_hash: cairo_felt::Felt252, _gas: &mut u128) -> SyscallResult<()> {
         println!("Called `replace_class({class_hash})` from MLIR.");
         Ok(())
     }
@@ -75,7 +75,7 @@ impl StarkNetSyscallHandler for SyscallHandler {
         class_hash: cairo_felt::Felt252,
         function_selector: cairo_felt::Felt252,
         calldata: &[cairo_felt::Felt252],
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<Vec<cairo_felt::Felt252>> {
         println!(
             "Called `library_call({class_hash}, {function_selector}, {calldata:?})` from MLIR."
@@ -88,7 +88,7 @@ impl StarkNetSyscallHandler for SyscallHandler {
         address: cairo_felt::Felt252,
         entry_point_selector: cairo_felt::Felt252,
         calldata: &[cairo_felt::Felt252],
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<Vec<cairo_felt::Felt252>> {
         println!(
             "Called `call_contract({address}, {entry_point_selector}, {calldata:?})` from MLIR."
@@ -100,7 +100,7 @@ impl StarkNetSyscallHandler for SyscallHandler {
         &mut self,
         address_domain: u32,
         address: cairo_felt::Felt252,
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<cairo_felt::Felt252> {
         println!("Called `storage_read({address_domain}, {address})` from MLIR.");
         Ok(address * &Felt252::new(3))
@@ -111,7 +111,7 @@ impl StarkNetSyscallHandler for SyscallHandler {
         address_domain: u32,
         address: cairo_felt::Felt252,
         value: cairo_felt::Felt252,
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<()> {
         println!("Called `storage_write({address_domain}, {address}, {value})` from MLIR.");
         Ok(())
@@ -121,7 +121,7 @@ impl StarkNetSyscallHandler for SyscallHandler {
         &mut self,
         keys: &[cairo_felt::Felt252],
         data: &[cairo_felt::Felt252],
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<()> {
         println!("Called `emit_event({keys:?}, {data:?})` from MLIR.");
         Ok(())
@@ -131,13 +131,17 @@ impl StarkNetSyscallHandler for SyscallHandler {
         &mut self,
         to_address: cairo_felt::Felt252,
         payload: &[cairo_felt::Felt252],
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<()> {
         println!("Called `send_message_to_l1({to_address}, {payload:?})` from MLIR.");
         Ok(())
     }
 
-    fn keccak(&self, input: &[u64], _gas: &mut u64) -> SyscallResult<cairo_native::starknet::U256> {
+    fn keccak(
+        &self,
+        input: &[u64],
+        _gas: &mut u128,
+    ) -> SyscallResult<cairo_native::starknet::U256> {
         println!("Called `keccak({input:?})` from MLIR.");
         Ok(U256(Felt252::from(1234567890).to_le_bytes()))
     }
@@ -146,7 +150,7 @@ impl StarkNetSyscallHandler for SyscallHandler {
         &self,
         _p0: cairo_native::starknet::Secp256k1Point,
         _p1: cairo_native::starknet::Secp256k1Point,
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<Option<cairo_native::starknet::Secp256k1Point>> {
         todo!()
     }
@@ -155,7 +159,7 @@ impl StarkNetSyscallHandler for SyscallHandler {
         &self,
         _x: cairo_native::starknet::U256,
         _y_parity: bool,
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<Option<cairo_native::starknet::Secp256k1Point>> {
         todo!()
     }
@@ -163,7 +167,7 @@ impl StarkNetSyscallHandler for SyscallHandler {
     fn secp256k1_get_xy(
         &self,
         _p: cairo_native::starknet::Secp256k1Point,
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<(cairo_native::starknet::U256, cairo_native::starknet::U256)> {
         todo!()
     }
@@ -172,7 +176,7 @@ impl StarkNetSyscallHandler for SyscallHandler {
         &self,
         _p: cairo_native::starknet::Secp256k1Point,
         _m: cairo_native::starknet::U256,
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<Option<cairo_native::starknet::Secp256k1Point>> {
         todo!()
     }
@@ -181,7 +185,7 @@ impl StarkNetSyscallHandler for SyscallHandler {
         &self,
         _x: cairo_native::starknet::U256,
         _y: cairo_native::starknet::U256,
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<Option<cairo_native::starknet::Secp256k1Point>> {
         todo!()
     }
@@ -190,7 +194,7 @@ impl StarkNetSyscallHandler for SyscallHandler {
         &self,
         _p0: cairo_native::starknet::Secp256k1Point,
         _p1: cairo_native::starknet::Secp256k1Point,
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<Option<cairo_native::starknet::Secp256k1Point>> {
         todo!()
     }
@@ -199,7 +203,7 @@ impl StarkNetSyscallHandler for SyscallHandler {
         &self,
         _x: cairo_native::starknet::U256,
         _y_parity: bool,
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<Option<cairo_native::starknet::Secp256k1Point>> {
         todo!()
     }
@@ -207,7 +211,7 @@ impl StarkNetSyscallHandler for SyscallHandler {
     fn secp256r1_get_xy(
         &self,
         _p: cairo_native::starknet::Secp256k1Point,
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<(cairo_native::starknet::U256, cairo_native::starknet::U256)> {
         todo!()
     }
@@ -216,7 +220,7 @@ impl StarkNetSyscallHandler for SyscallHandler {
         &self,
         _p: cairo_native::starknet::Secp256k1Point,
         _m: cairo_native::starknet::U256,
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<Option<cairo_native::starknet::Secp256k1Point>> {
         todo!()
     }
@@ -225,7 +229,7 @@ impl StarkNetSyscallHandler for SyscallHandler {
         &self,
         _x: cairo_native::starknet::U256,
         _y: cairo_native::starknet::U256,
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<Option<cairo_native::starknet::Secp256k1Point>> {
         todo!()
     }
